chore(suite): rename WalletLayoutHeader
