+++ packages/suite/src/views/wallet/anonymize/index.tsx
@@ -1,5 +1,5 @@
 import { useSelector } from 'src/hooks/suite';
-import { WalletLayout, WalletLayoutHeader } from 'src/components/wallet';
+import { WalletLayout, WalletSubpageHeading } from 'src/components/wallet';
 import { CoinjoinConfirmation } from 'src/views/wallet/anonymize/components/CoinjoinConfirmation';
 
 const Anonymize = () => {
@@ -9,7 +9,7 @@ const Anonymize = () => {
         <WalletLayout title="TR_NAV_ANONYMIZE" isSubpage account={selectedAccount}>
             {selectedAccount.status === 'loaded' && (
                 <>
-                    <WalletLayoutHeader title="TR_NAV_ANONYMIZE" />
+                    <WalletSubpageHeading title="TR_NAV_ANONYMIZE" />
                     <CoinjoinConfirmation account={selectedAccount.account} />
                 </>
             )}
